Feature/nhh comment&&qa

+++ clone/src/main/java/com/shopee/clone/service/promotion/IPromotionService.java
@@ -10,12 +10,12 @@ import java.util.List;
 public interface IPromotionService {
     ResponseEntity<?> createPromotion(PromotionRequestCreate promotionRequestCreate);
     ResponseEntity<?> editStatusPromotion(Long promotionId, Boolean status);
-    Boolean isValidPromotion(String name, Integer purchasedAmount);
+    Boolean isValidPromotion(String name, Double purchasedAmount);
     ResponseEntity<?> addPromotionBeLongUser(Long userId, List<Long> listPromotionId);
     //Call To this Service is Enough
-    Boolean checkValidUsage(Long userId, String promotionName, Integer purchasedAmount);
-    Boolean minusUsage(Long userId, String promotionName, Integer purchasedAmount);
-    Boolean plusUsage(Long userId, String promotionName, Integer purchasedAmount);
+    Boolean checkValidUsage(Long userId, String promotionName, Double purchasedAmount);
+    Boolean minusUsage(Long userId, String promotionName, Double purchasedAmount);
+    Boolean plusUsage(Long userId, String promotionName, Double purchasedAmount);
     ResponseEntity<?> getAllPromotionAvailable(Long userId);
     ResponseEntity<?> getAllPromotionBySellerId(Long sellerId);
     ResponseEntity<?> getPromotionOfUser(Long userId);
